dist-cc: auto ref, print changes, uint32 -> 64 bit

+++ dist_apps/cc_push-filter/gen_cuda.cuh
@@ -7,8 +7,8 @@
 #include "Galois/Runtime/Cuda/cuda_helpers.h"
 
 struct CUDA_Context : public CUDA_Context_Common {
-	struct CUDA_Context_Field<uint32_t> comp_current;
-	struct CUDA_Context_Field<uint32_t> comp_old;
+	struct CUDA_Context_Field<unsigned long long> comp_current;
+	struct CUDA_Context_Field<unsigned long long> comp_old;
 };
 
 struct CUDA_Context *get_CUDA_context(int id) {
@@ -50,23 +50,23 @@ void bitset_comp_current_reset_cuda(struct CUDA_Context *ctx, size_t begin, size
   reset_bitset_field(&ctx->comp_current, begin, end);
 }
 
-uint32_t get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID) {
-	uint32_t *comp_current = ctx->comp_current.data.cpu_rd_ptr();
+unsigned long long get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID) {
+	unsigned long long *comp_current = ctx->comp_current.data.cpu_rd_ptr();
 	return comp_current[LID];
 }
 
-void set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+void set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	comp_current[LID] = v;
 }
 
-void add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+void add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	comp_current[LID] += v;
 }
 
-bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_current = ctx->comp_current.data.cpu_wr_ptr();
+bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_current = ctx->comp_current.data.cpu_wr_ptr();
 	if (comp_current[LID] > v) {
 		comp_current[LID] = v;
     return true;
@@ -74,67 +74,67 @@ bool min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t
   return false;
 }
 
-void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
-	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_current, from_id, v);
+void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
+	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_current, from_id, v);
 }
 
-void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
-	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_current, from_id, v);
+void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
+	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_current, from_id, v);
 }
 
-void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v, uint32_t i) {
-	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_current, from_id, v, i);
+void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v, uint32_t i) {
+	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_current, from_id, v, i);
 }
 
-void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode, uint32_t i) {
-	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
+void batch_get_reset_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode, uint32_t i) {
+	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
 }
 
-void batch_set_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMirror, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_mirror_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMirror, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, setOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, addOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_add_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, addOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, minOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_min_node_comp_current_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, minOp>(ctx, &ctx->comp_current, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
 void bitset_comp_old_clear_cuda(struct CUDA_Context *ctx) {
 	ctx->comp_old.is_updated.cpu_rd_ptr()->reset();
 }
 
-uint32_t get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID) {
-	uint32_t *comp_old = ctx->comp_old.data.cpu_rd_ptr();
+unsigned long long get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID) {
+	unsigned long long *comp_old = ctx->comp_old.data.cpu_rd_ptr();
 	return comp_old[LID];
 }
 
-void set_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_old = ctx->comp_old.data.cpu_wr_ptr();
+void set_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_old = ctx->comp_old.data.cpu_wr_ptr();
 	comp_old[LID] = v;
 }
 
-void add_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_old = ctx->comp_old.data.cpu_wr_ptr();
+void add_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_old = ctx->comp_old.data.cpu_wr_ptr();
 	comp_old[LID] += v;
 }
 
-bool min_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) {
-	uint32_t *comp_old = ctx->comp_old.data.cpu_wr_ptr();
+bool min_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, unsigned long long v) {
+	unsigned long long *comp_old = ctx->comp_old.data.cpu_wr_ptr();
 	if (comp_old[LID] > v) {
 		comp_old[LID] = v;
     return true;
@@ -142,43 +142,43 @@ bool min_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned LID, uint32_t v) 
   return false;
 }
 
-void batch_get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
-	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_old, from_id, v);
+void batch_get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
+	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_old, from_id, v);
 }
 
-void batch_get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<uint32_t, sharedMaster, false>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<unsigned long long, sharedMaster, false>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v) {
-	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_old, from_id, v);
+void batch_get_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v) {
+	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_old, from_id, v);
 }
 
-void batch_get_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode) {
-	batch_get_shared_field<uint32_t, sharedMirror, false>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_get_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode) {
+	batch_get_shared_field<unsigned long long, sharedMirror, false>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_get_reset_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, uint32_t *v, uint32_t i) {
-	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_old, from_id, v, i);
+void batch_get_reset_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long *v, uint32_t i) {
+	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_old, from_id, v, i);
 }
 
-void batch_get_reset_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t *v_size, DataCommMode *data_mode, uint32_t i) {
-	batch_get_shared_field<uint32_t, sharedMirror, true>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
+void batch_get_reset_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t *v_size, DataCommMode *data_mode, unsigned long i) {
+	batch_get_shared_field<unsigned long long, sharedMirror, true>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode, i);
 }
 
-void batch_set_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMirror, setOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_mirror_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMirror, setOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_set_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, setOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_set_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, setOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_add_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, addOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_add_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, addOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
 
-void batch_min_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, uint32_t *v, size_t v_size, DataCommMode data_mode) {
-	batch_set_shared_field<uint32_t, sharedMaster, minOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
+void batch_min_node_comp_old_cuda(struct CUDA_Context *ctx, unsigned from_id, unsigned long long int *bitset_comm, unsigned int *offsets, unsigned long long *v, size_t v_size, DataCommMode data_mode) {
+	batch_set_shared_field<unsigned long long, sharedMaster, minOp>(ctx, &ctx->comp_old, from_id, bitset_comm, offsets, v, v_size, data_mode);
 }
